+++ src/Components/Redirectable/RedirectableMoveBlock.cs
@@ -1,6 +1,5 @@
-﻿#nullable enable
+#nullable enable
 
-using Celeste;
 using Celeste.Mod.CommunalHelper.Entities;
 using MonoMod.Utils;
 using System.Collections;
@@ -11,21 +10,8 @@ using Directions = Celeste.MoveBlock.Directions;
 
 namespace Celeste.Mod.CommunalHelper.Components;
 
-/// <summary>
-/// get/set
-///     homeAngle
-///     direction
-///     targetSpeed
-///     speed
-/// get
-///     canSteer
-///     moveSfx
-/// set
-///     targetAngle
-///     angle
-/// </summary>
 [Tracked(true)]
-public class Redirectable : Component
+public class RedirectableMoveBlock : Redirectable
 {
     // Pre-initialize this with some known types
     private static readonly Dictionary<Type, FieldInfo> reflectionCache = new()
@@ -36,9 +22,21 @@ public class Redirectable : Component
         { typeof(MoveSwapBlock), typeof(MoveSwapBlock).GetNestedTypes(BindingFlags.NonPublic).First(t => t.Name.StartsWith("<Controller>")).GetField("<>4__this") },
     };
 
+    public DynamicData Data;
+    private Action<Coroutine> onResumeAction;
+    private Action<Coroutine> onBreakAction;
+    public RedirectableMoveBlock(DynamicData Data) : base()
+    {
+        this.Data = Data;
+
+        onResumeAction = GetControllerDelegate(Data, 3);
+        onBreakAction = GetControllerDelegate(Data, 4);
+    }
+
     public Func<float>? Get_Speed = null;
     public Action<float>? Set_Speed = null;
-    public float Speed
+
+    public override float Speed
     {
         get => Get_Speed?.Invoke() ?? Data.Get<float>("speed");
         set
@@ -50,9 +48,10 @@ public class Redirectable : Component
         }
     }
 
+
     public Func<float>? Get_TargetSpeed = null;
     public Action<float>? Set_TargetSpeed = null;
-    public float TargetSpeed
+    public override float TargetSpeed
     {
         get => Get_TargetSpeed?.Invoke() ?? Data.Get<float>("targetSpeed");
         set
@@ -66,7 +65,8 @@ public class Redirectable : Component
 
     public Func<Directions>? Get_Direction = null;
     public Action<Directions>? Set_Direction = null;
-    public Directions Direction
+
+    public override Directions Direction
     {
         get => Get_Direction?.Invoke() ?? Data.Get<Directions>("direction");
         set
@@ -78,68 +78,22 @@ public class Redirectable : Component
         }
     }
 
-    public Func<float>? Get_HomeAngle = null;
-    public Action<float>? Set_HomeAngle = null;
-    public float HomeAngle
+
+    public override float Angle
     {
-        get => Get_HomeAngle?.Invoke() ?? Data.Get<float>("homeAngle");
-        set
+        get => Data.Get<float>("Angle"); set
         {
-            if (Set_HomeAngle != null)
-                Set_HomeAngle.Invoke(value);
-            else
-                Data.Set("homeAngle", value);
+            Data.Set("angle", value);
+            Data.Set("homeAngle", value);
+            Data.Set("targetAngle", value);
         }
     }
 
     public Func<bool>? Get_CanSteer = null;
-    public bool CanSteer => Get_CanSteer?.Invoke() ?? Data.Get<bool>("canSteer");
+    public override bool CanSteer => Get_CanSteer?.Invoke() ?? Data.Get<bool>("canSteer");
 
     public Func<SoundSource>? Get_MoveSfx = null;
     public SoundSource MoveSfx => Get_MoveSfx?.Invoke() ?? Data.Get<SoundSource>("moveSfx");
-
-    public Action<float>? Set_TargetAngle = null;
-    public float TargetAngle
-    {
-        set
-        {
-            if (Set_TargetAngle != null)
-                Set_TargetAngle(value);
-            else
-                Data.Set("targetAngle", value);
-        }
-    }
-
-    public Action<float>? Set_Angle = null;
-    public float Angle
-    {
-        set
-        {
-            if (Set_Angle != null)
-                Set_Angle(value);
-            else
-                Data.Set("angle", value);
-        }
-    }
-
-    public float? InitialAngle;
-    public Directions? InitialDirection;
-
-    public Action<float>? OnStartShaking;
-    public Action<Vector2>? OnMoveTo;
-
-    public Action<Coroutine> OnResume;
-    public Action<Coroutine> OnBreak;
-
-    public DynamicData Data;
-
-    public Redirectable(DynamicData entityData, Action<Coroutine>? onResume = null, Action<Coroutine>? onBreak = null)
-        : base(false, false)
-    {
-        Data = entityData;
-        OnResume = onResume ?? GetControllerDelegate(entityData, 3);
-        OnBreak = onBreak ?? GetControllerDelegate(entityData, 4);
-    }
 
     public static Action<Coroutine> GetControllerDelegate(DynamicData targetData, int jumpPoint)
     {
@@ -161,28 +115,43 @@ public class Redirectable : Component
         };
     }
 
-    public void StartShaking(float time = 0)
-    {
-        if (OnStartShaking != null)
-            OnStartShaking.Invoke(time);
-        else
-            (Entity as Platform)?.StartShaking(time);
-    }
-
-    public void MoveTo(Vector2 position)
-    {
-        if (OnMoveTo != null)
-            OnMoveTo.Invoke(position);
-        else
-            (Entity as Platform)?.MoveTo(position);
-    }
 
     public void ResetBlock()
     {
-        if (InitialAngle != null)
-            Angle = TargetAngle = HomeAngle = InitialAngle.Value;
-        if (InitialDirection != null)
-            Direction = InitialDirection.Value;
+        Angle = InitialAngle;
+        Direction = InitialDirection;
+    }
+
+    public override void MoveTo(Vector2 to)
+    {
+        (Entity as Platform)?.MoveTo(to);
+    }
+
+    public override void OnPause(Coroutine moveCoroutine, float shakeTime)
+    {
+        MoveSfx?.Param("redirect_slowdown", 1f);
+        (Entity as Platform)?.StartShaking(shakeTime);
+    }
+
+    public override void OnResume(Coroutine moveCoroutine)
+    {
+        onResumeAction(moveCoroutine);
+    }
+
+    public override void OnBreak(Coroutine moveCoroutine)
+    {
+        onBreakAction(moveCoroutine);
+    }
+
+    public override void BeforeBreakEffect()
+    {
+        MoveSfx?.Stop();
+    }
+
+    public override void OnRedirectEffect(float shakeTime)
+    {
+        (Entity as Platform)?.StartShaking(shakeTime);
+        MoveSfx?.Param("redirect_slowdown", 0f);
     }
 
     internal static void Load()
@@ -200,13 +169,22 @@ public class Redirectable : Component
     {
         orig(self, position, width, height, direction, canSteer, fast);
         if (self.GetType() == typeof(MoveBlock))
-            self.Add(new Redirectable(new DynamicData(typeof(MoveBlock), self)));
+            self.Add(new RedirectableMoveBlock(new DynamicData(typeof(MoveBlock), self)));
     }
 
     private static void MoveBlock_BreakParticles(On.Celeste.MoveBlock.orig_BreakParticles orig, MoveBlock self)
     {
         orig(self);
-        self.Get<Redirectable>()?.ResetBlock();
+        ((RedirectableMoveBlock) self.Get<Redirectable>())?.ResetBlock();
     }
 
+    protected override float GetInitialAngle()
+    {
+        return Data.Get<float>("homeAngle");
+    }
+
+    protected override Directions GetInitialDirection()
+    {
+        return Direction;
+    }
 }
